replate hyerpband.py by asha.py
